New algo + HTML output
I've changed the if statement tree to reflect a more strict structure. This is to rate a specific extension at the moment and I need not worry about fringe cases...... at the moment.
Also, the csv-rater.py will take an argument of which ini file to read so you can have multiple ones parsing multiple configs. This will help in the cronjob Ima use.